record 타입 변환
