+++ .idea/workspace.xml
@@ -1,0 +1,113 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="1fe1570b-38b0-4a5d-8519-5916e0d6f809" name="변경" comment="">
+      <change beforePath="$PROJECT_DIR$/namyeongjun/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/.gitignore" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/DemoApplication.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/DemoApplication.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/controller/RepeatController.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/controller/RepeatController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/dto/ErrorResponseDto.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/dto/ErrorResponseDto.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/dto/RepeatRequestDto.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/dto/RepeatRequestDto.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/dto/RepeatResponseDto.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/dto/RepeatResponseDto.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/java/com/example/demo/service/RepeatService.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/java/com/example/demo/service/RepeatService.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/main/resources/application.properties" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/main/resources/application.properties" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/namyeongjun/test/java/com/example/demo/DemoApplicationTests.java" beforeDir="false" afterPath="$PROJECT_DIR$/namyeongjun/src/test/java/com/example/demo/DemoApplicationTests.java" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="ExternalProjectsData">
+    <projectState path="$PROJECT_DIR$">
+      <ProjectState />
+    </projectState>
+  </component>
+  <component name="ExternalProjectsManager">
+    <system id="GRADLE">
+      <state>
+        <projects_view>
+          <tree_state>
+            <expand />
+            <select />
+          </tree_state>
+        </projects_view>
+      </state>
+    </system>
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 0
+}</component>
+  <component name="ProjectId" id="3BAd1Y4Mc3qfQGHtdXKLoMUexzr" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;Gradle.7th-BE-Blog 빌드.executor&quot;: &quot;Run&quot;,
+    &quot;Gradle.C:/Users/skaud/7th-BE-Blog/settings.gradle.executor&quot;: &quot;Run&quot;,
+    &quot;ModuleVcsDetector.initialDetectionPerformed&quot;: &quot;true&quot;,
+    &quot;RequestMappingsPanelOrder0&quot;: &quot;0&quot;,
+    &quot;RequestMappingsPanelOrder1&quot;: &quot;1&quot;,
+    &quot;RequestMappingsPanelWidth0&quot;: &quot;75&quot;,
+    &quot;RequestMappingsPanelWidth1&quot;: &quot;75&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.TerminalTabsStorage.copyFrom.TerminalArrangementManager.252&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.git.unshallow&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.typescript.service.memoryLimit.init&quot;: &quot;true&quot;,
+    &quot;Spring Boot.DemoApplication.executor&quot;: &quot;Run&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;나명준/1주차&quot;,
+    &quot;ignore.virus.scanning.warn.message&quot;: &quot;true&quot;,
+    &quot;kotlin-language-version-configured&quot;: &quot;true&quot;,
+    &quot;last_opened_file_path&quot;: &quot;C:/Users/skaud/7th-BE-Blog&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
+  }
+}</component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\skaud\7th-BE-Blog\namyeongjun\src" />
+      <recent name="C:\Users\skaud\7th-BE-Blog\namyeongjun" />
+    </key>
+  </component>
+  <component name="RunManager">
+    <configuration name="DemoApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
+      <module name="7th-BE-Blog" />
+      <option name="SPRING_BOOT_MAIN_CLASS" value="com.example.demo.DemoApplication" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+  </component>
+  <component name="SharedIndexes">
+    <attachedChunks>
+      <set>
+        <option value="bundled-jdk-30f59d01ecdd-2fc7cc6b9a17-intellij.indexing.shared.core-IU-253.30387.90" />
+        <option value="bundled-js-predefined-d6986cc7102b-9b0f141eb926-JavaScript-IU-253.30387.90" />
+      </set>
+    </attachedChunks>
+  </component>
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="디폴트 작업">
+      <changelist id="1fe1570b-38b0-4a5d-8519-5916e0d6f809" name="변경" comment="" />
+      <created>1773935593667</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1773935593667</updated>
+      <workItem from="1773935594836" duration="24032000" />
+    </task>
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="com.intellij.coverage.CoverageDataManagerImpl">
+    <SUITE FILE_PATH="coverage/7th_BE_Blog$DemoApplication.ic" NAME="DemoApplication 커버리지 결과" MODIFIED="1774346603786" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="idea" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="true" />
+  </component>
+</project>
