+++ namyeongjun/.postman/resources.yaml
@@ -1,0 +1,9 @@
+# Use this workspace to collaborate
+workspace:
+  id: 695def32-7347-4f12-844d-57e3a17a468b
+
+# All resources in the `postman/` folder are automatically registered in Local View.
+# Point to additional files outside the `postman/` folder to register them individually. Example:
+#localResources:
+#  collections:
+#    - ../tests/E2E Test Collection/
