+++ namyeongjun/src/main/java/com/example/demo/dto/RepeatRequestDto.java
@@ -1,0 +1,8 @@
+package com.example.demo.dto;
+
+
+import jakarta.validation.constraints.NotBlank;
+
+
+
+public record RepeatRequestDto (@NotBlank(message = "내용을 입력해주세요.") String value){ }
