+++ namyeongjun/src/main/java/com/example/demo/dto/RepeatResponseDto.java
@@ -1,0 +1,6 @@
+package com.example.demo.dto;
+
+
+
+
+public record RepeatResponseDto(String string_one,String string_two) {}
